Allow disabling of violation logging
When some tasks take too long, a violation is logged using the default
logging package.  We don't want to add more dependencies to this
package, but we also don't want to log these violations in every
circumstance (e.g. when used in a CLI).

Allows the caller enabled a tracing mutex to specify they want to
Disable violation logging.